Add memories package, moving MemoriesManager to its own module
